fix: Align implementation
